+++ NEWS.md
@@ -2,8 +2,10 @@
 
 ## New Features
 
-* `update_glance` now uses `glue()` for `title`, `text`, and `subtext` values
-* `upadte_glance` coerces non-integer `count` and `percent` values to integer rather than aborting
+* `pushover()`: `glue()` is now evaluated prior to character limit checks
+* `update_glance`
+  * `glue()` formatting for `title`, `text`, and `subtext` values
+  * Non-integer `count` and `percent` values are now coerced to integer rather than aborting
 
 
 # pushoverr 1.1.0 (2021-11-13)